requirements.txt
fastapi
uvicorn[standard]
beanie
motor
pydantic[email]
httpx
python-dotenv
pydantic-settings
openai
scikit-learn
scipy
numpy
